requirements.txt
pyautogui>=0.9.54
Pillow>=10.0.0
opencv-python>=4.9.0
pyobjc-framework-Quartz>=10.0
